+++ lib/glimmer/libui/shape/line.rb
@@ -53,6 +53,13 @@ module Glimmer
           !parent.is_a?(Figure) && end_x && end_y
         end
         
+        def move_by(x_delta, y_delta)
+          self.x += x_delta
+          self.y += y_delta
+          self.end_x += x_delta
+          self.end_y += y_delta
+        end
+        
         def perfect_shape
           perfect_shape_dependencies = [x, y, end_x, end_y]
           if perfect_shape_dependencies != @perfect_shape_dependencies