+++ lib/glimmer/libui/shape/rectangle.rb
@@ -33,6 +33,11 @@ module Glimmer
           super
         end
         
+        def move_by(x_delta, y_delta)
+          self.x += x_delta
+          self.y += y_delta
+        end
+        
         def perfect_shape
           perfect_shape_dependencies = [x, y, width, height]
           if perfect_shape_dependencies != @perfect_shape_dependencies